pipeline-manager: distinguish two pipeline API categories
Distinguishes two categories for the pipeline API:
- Management: manage the pipeline through the database
- Interaction: interact with the deployed pipeline

Part of the initiative to gradually improve the API.
